fix(p2p,consensus): address PR #264 review feedback
Addresses the 7 blocking + 10 important + 6 nit items raised in review.
Landing plan lives in tasks/pr264-fixes.md.

Blocking — consensus semantics and correctness:
  * Round duration: StartTime reset to wall clock; Duration uses
    time.Since to keep the pair balanced (engine.go:186,1237).
  * Replay-delta apply divergence no longer charges the peer — engine
    bug, not peer misbehavior (router.go:576); log escalated to ERROR
    with "ENGINE DIVERGENCE" marker.
  * Self-originated broadcasts skip the squelch filter; relays apply
    it AND exclude the originating peer. Overlay.BroadcastFromValidator
    renamed to RelayFromValidator; new RelayValidation method; rippled
    OverlayImpl.cpp:1133-1137,1158-1163.
  * OnProposal/OnValidation plumb originPeer through the engine so
    relay forwards can exclude the origin. Engine.Adaptor gains
    RelayValidation + UpdateRelaySlot.
  * Router calls UpdateRelaySlot on every trusted inbound proposal +
    validation — feeds the Relay.OnMessage loop that drives mtSQUELCH
    emission. Previously only called from tests. Rippled PeerImp.cpp:
    1737, 2385, 3013, 3049.
  * Unified ledger-acquisition dedup: isAcquiring() checks BOTH the
    replayer in-flight map AND inboundLedger. Eliminates the race
    where two status changes at the same seq armed both paths.
    Rippled InboundLedgers::find.
  * ModeSwitchedLedger restored: post-handleWrongLedger and
    post-OnLedger startRoundLocked(recovering=true) suppresses
    proposal emission AND validation emission for one round, matching
    rippled Consensus.h:1107,1457. NEXT round auto-promotes back to
    ModeProposing.
  * checkLedger switches on validation SUPPORT (trusted count) rather
    than full-validation quorum. Uses ValidationTracker.GetTrustedSupport
    for netLgr vs ourID comparison. Prevents catch-up stranding.
    Rippled vals.getPreferred() RCLConsensus.cpp:300-303.

Important — subsystem hardening:
  * maintenanceTick reaps stuck inbound-ledger acquisitions alongside
    replayer timeouts.
  * Ingress mtREPLAY_DELTA_REQ + mtPROOF_PATH_REQ gated on handshake
    ledgerreplay feature; unnegotiated peers charged + dropped.
    Fixed latent X-Protocol-Ctl header never being written on outbound
    handshake requests (WriteRawHandshakeRequest whitelist).
  * Oversize replay-delta response returns reNO_LEDGER (not
    reBAD_REQUEST) to avoid charging the requester feeMalformedRequest.
  * replayDeltaTimeout comment + Apply ShouldRetry comment rewritten
    honestly — no false "rippled parity" claims.
  * ValidationTracker: Full=true filter, seq floor via SetMinSeq,
    negative-UNL filter via SetNegativeUNL, trusted-set reload on every
    acceptLedger (amendment / UNL changes). Rippled LedgerMaster.cpp:
    886,952.
  * STValidation: sfValidatedHash emission infrastructure — struct
    fields, serializer, signer-side preimage, parser, and adaptor hook
    GetValidatedLedgerHash. sfAmendments encoder/decoder wired so a
    future amendment-voting policy can populate Amendments without
    further wire-format work. Rippled RCLConsensus.cpp:853-894.
  * peerLCLs cleaned on disconnect (Overlay.SetPeerDisconnectCallback
    + Router.HandlePeerDisconnect). getNetworkLedger dedupes
    peerLCL votes against already-counted trusted-proposer hashes so a
    validator connected as a peer doesn't double-vote.
  * Bad-data counter → signed int64 balance with per-reason weights
    (feeInvalidData=400, feeMalformedRequest=200, feeRequestNoReply=10,
    default=100) and exponential decay halving every 10s. Eviction
    threshold lifted to 1000 to match the new scale. Rippled
    Resource::Fees.cpp:26-43.

Nits:
  * extractTransactionIndex: streaming field-header walk replaces
    full-decode — O(n) skip until sfTransactionIndex.
  * DefaultMaxInFlightReplays comment honest about non-parity with
    rippled's MAX_TASKS.
  * Slot.DeletePeer resets Count=0/LastMessage=now on the deleted
    peer regardless of branch, matching rippled Slot.h:479-480.
  * Removed unused PeerCapabilities fields (ProtocolMajor/Minor,
    NetworkID, ListeningPort, SupportsCrawl, IsValidator) — none
    were populated.
  * Dropped-message atomic counters (Overlay.DroppedMessages,
    DroppedLedgerResponses) replace silent default: drops.
  * pendingValidation LRU drop logs at warn — previously silent event
    loss for the oldest hash.

Cleanup:
  * gofmt fixes across 3 touched files after imports changed.
  * unconvert: removed redundant int64() wrap in p2p integration test.

+++ internal/testing/p2p/replay_delta_apply_integration_test.go
@@ -53,7 +53,7 @@ func TestReplayDelta_Apply_Integration(t *testing.T) {
 	// Build + sign a real Payment alice → bob using the env's signer.
 	aliceSeq := env.Seq(alice)
 	pay := payment.NewPayment(alice.Address, bob.Address,
-		tx.NewXRPAmount(int64(xrplgoTesting.XRP(123))))
+		tx.NewXRPAmount(xrplgoTesting.XRP(123)))
 	pay.Sequence = &aliceSeq
 	pay.Fee = "10"
 	env.SignWith(pay, alice)
@@ -142,7 +142,7 @@ func TestReplayDelta_Apply_DivergenceFromTef(t *testing.T) {
 
 	aliceSeq := env.Seq(alice)
 	pay := payment.NewPayment(alice.Address, bob.Address,
-		tx.NewXRPAmount(int64(xrplgoTesting.XRP(123))))
+		tx.NewXRPAmount(xrplgoTesting.XRP(123)))
 	pay.Sequence = &aliceSeq
 	pay.Fee = "10"
 	env.SignWith(pay, alice)
